refactor: header menubar, footer 제거
