--- components/Layout/Footer.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-const Footer = () => {
-  return <div className="py-10 bg-gray-800 text-center">Footer</div>
-}
-
-export default Footer
